+++ src/Components/TagList/TagList.module.less
@@ -66,3 +66,7 @@
 .contact-filter {
     width: 100%;
 }
+
+.subscriptionList {
+    width: 100%;
+}
