+++ src/Components/TriggerInfo/Components/TriggerDetails/TriggerDetails.module.less
@@ -1,0 +1,48 @@
+.info-section {
+    display: grid;
+    grid-auto-columns: 1fr 288px;
+}
+
+.list {
+    display: flex;
+    flex-wrap: wrap;
+    margin-top: 0;
+    margin-bottom: 0;
+    overflow: hidden;
+    grid-column: 1;
+    dt {
+        font-style: italic;
+        margin-top: 10px;
+        width: 150px;
+        color: @colorGrey;
+    }
+    dd {
+        margin-top: 10px;
+        margin-left: 0;
+        width: calc(~"100% - 150px");
+        &.description {
+            word-wrap: break-word;
+            overflow: hidden;
+        }
+        &.codeEditor {
+            margin-left: -16px;
+        }
+    }
+}
+
+.copyButtonWrapper {
+    position: relative;
+}
+
+.copyButton {
+    position: absolute;
+    top: 5px;
+    left: -24px;
+}
+
+.exception-explanation {
+    white-space: pre-wrap;
+    .line-2 {
+        margin-left: 20px;
+    }
+}
