+++ src/Components/TriggerInfo/Components/TriggerHeader/TriggerHeader.module.less
@@ -1,0 +1,23 @@
+.header {
+    display: flex;
+    align-items: baseline;
+}
+
+.title {
+    margin-top: 0;
+    margin-bottom: 15px;
+    flex-grow: 1;
+    font-weight: 300;
+    font-size: 32px;
+    line-height: 40px;
+}
+
+.editLink {
+    padding-right: 24px;
+    white-space: nowrap;
+}
+
+.controls {
+    display: flex;
+    align-items: baseline;
+}
