+++ src/Components/TriggerInfo/TriggerInfo.module.less
@@ -1,76 +1,12 @@
-.header {
-    display: flex;
-    align-items: baseline;
-}
-.title {
-    margin-top: 0;
-    margin-bottom: 15px;
-    flex-grow: 1;
-    font-weight: 300;
-    font-size: 32px;
-    line-height: 40px;
-}
-.controls {
-    display: flex;
-    align-items: baseline;
-}
-
-.list {
-    display: flex;
-    flex-wrap: wrap;
-    margin-top: 0;
-    margin-bottom: 0;
-    overflow: hidden;
-    grid-column: 1;
-    dt {
-        font-style: italic;
-        margin-top: 10px;
-        width: 150px;
-        color: @colorGrey;
-    }
-    dd {
-        margin-top: 10px;
-        margin-left: 0;
-        width: calc(~"100% - 150px");
-        &.description {
-            word-wrap: break-word;
-            overflow: hidden;
-        }
-        &.codeEditor {
-            margin-left: -16px;
-        }
-    }
-}
-.copyButtonWrapper {
-    position: relative;
-}
-.copyButton {
-    position: absolute;
-    top: 5px;
-    left: -24px;
-}
-.maintenance-info {
-    color: @linkColor;
-}
-.exception-explanation {
-    white-space: pre-wrap;
-    .line-2 {
-        margin-left: 20px;
-    }
-}
 .info-section {
     display: grid;
     grid-auto-columns: 1fr 288px;
 }
+
 .state-chart {
     grid-column: 2;
     padding-left: 10px;
     min-width: 288px;
     display: flex;
     align-items: center;
-}
-
-.editLink {
-    padding-right: 24px;
-    white-space: nowrap;
 }
